BOL-375 refactor: Mapper 서버 맞춰서 리팩터링
